Fix loss_utils validation order, get_one_hot dtype, and redundant cast
- check_loss_inputs: move 5D guard before channel checks so a 4D input
  with 1 prediction channel raises 'must be 5D' rather than the misleading
  'number of classes' error
- get_one_hot: return float32 instead of casting to int8; removes silent
  type promotion in every downstream loss arithmetic operation
- SegmentationLoss.preprocess: drop now-redundant .to(torch.float32) call
- SoftSkeletonize: replace old-style super(SoftSkeletonize, self) with super()
- Add test asserting the 5D check fires before the channel check

+++ mist/loss_functions/base.py
@@ -76,7 +76,7 @@ class SegmentationLoss(nn.Module, ABC):
             y_true = y_true[:, 1:]
             y_pred = y_pred[:, 1:]
 
-        return y_true.to(torch.float32), y_pred
+        return y_true, y_pred
 
     @abstractmethod
     def forward(
